requirements.txt
django-crontab
git+https://github.com/manahl/arctic.git
vectorbt
django-admin